refactor(football): split transfermarkt into client/parser/service package
Break the 686-line monolith into three focused modules:
- parser.py: HTML parsing (BeautifulSoup → dataclasses), regex constants
- client.py: async HTTP fetch operations
- service.py: public facade with POSITION_ROLES and market-value utils
